+++ Sources/VComponents/Presentation/Components + Architectural Pattern Helpers/Spinner/VSpinnerParameters.swift
@@ -1,5 +1,5 @@
 //
-//  VSpinnerParameters.swift
+//  VContinuousSpinnerParameters.swift
 //  VComponents
 //
 //  Created by Vakhtang Kontridze on 02.10.22.
@@ -7,19 +7,22 @@
 
 import SwiftUI
 
-/// Parameters for presenting an `VSpinner`.
+/// Parameters for presenting an `VContinuousSpinner`.
 ///
-///     @State private var parameters: VSpinnerParameters = .init()
+///     @State private var parameters: VContinuousSpinnerParameters = .init()
 ///
 ///     var body: some View {
 ///         content
 ///             .vContinuousSpinner(parameters: parameters)
 ///     }
 ///
-public struct VSpinnerParameters {
+public struct VContinuousSpinnerParameters {
     // MARK: Properties
     /// Appearance.
     public var appearance: VContinuousSpinnerAppearance
+    
+    /// Appearance delay.
+    public var appearanceDelay: Duration?
     
     /// Indicates if interaction is enabled.
     public var isInteractionEnabled: Bool
@@ -28,13 +31,15 @@ public struct VSpinnerParameters {
     public var attributes: [String: Any]
 
     // MARK: Initializers
-    /// Initializes `VSpinnerParameters`.
+    /// Initializes `VContinuousSpinnerParameters`.
     public init(
         appearance: VContinuousSpinnerAppearance = .init(),
+        appearanceDelay: Duration? = nil,
         isInteractionEnabled: Bool = true,
         attributes: [String: Any] = [:]
     ) {
         self.appearance = appearance
+        self.appearanceDelay = appearanceDelay
         self.isInteractionEnabled = isInteractionEnabled
         self.attributes = attributes
     }