+++ CHANGELOG.md
@@ -19,6 +19,10 @@ VCodeEntryView
 - State now support `empty` and `filled` cases
 - Fixed and issue where underlying `TextField` became visible with semi-transparent character background color
 
+Presentation
+
+- `appearanceDelay` is added to `VContinuousSpinnerParameters`
+
 ### [8.0.1(78)](https://github.com/VakhoKontridze/VComponents/releases/tag/8.0.1) — 2026 04 24
 
 General